Don't translate generic default trait methods.
I think Self was in the fn's type_scheme.generics.types in the old nightly.

This fixes the 3 examples exhibiting the #40 behavior, so I also added the default methods back to the tests that were created from those examples.

(The cmp example and test are still failing, but it's surely because of substs/monomorphization problems)